Move methods to AveragePriceUtil
